feat: rename index to cluster
