+++ src/commands/gameplay_primary/scavenge.ts
@@ -1,6 +1,6 @@
 import { AsyncQueue } from '@sapphire/async-queue';
 import { ActionRowBuilder, ButtonBuilder, ButtonInteraction, ButtonStyle, ChatInputCommandInteraction, ComponentType, EmbedBuilder, InteractionResponse, Message, SlashCommandBuilder } from 'discord.js';
-import { speciesInfo } from '../..';
+import { speciesInfo } from '../../cluster';
 import Quid from '../../models/quid';
 import QuidToServer from '../../models/quidToServer';
 import Server from '../../models/server';
