+++ src/events/guildCreate.ts
@@ -1,5 +1,5 @@
 import { Guild } from 'discord.js';
-import { client } from '../index';
+import { client } from '../cluster';
 import { DiscordEvent } from '../typings/main';
 import { createGuild } from '../utils/updateGuild';
